requirements.txt
altgraph==0.17.2
asgiref==3.5.2
Django==4.0.5
django-ckeditor==6.4.2
django-js-asset==2.0.0
future==0.18.2
pefile==2022.5.30
Pillow==9.2.0
pyinstaller-hooks-contrib==2022.7
pywin32-ctypes==0.2.0
sqlparse==0.4.2
tzdata==2022.1
